FrameRate + bajo, player movido
Hay que solucionar el problema del flickering

+++ MonkeyDelivery/Src/Control/States/MenuState.cpp
@@ -42,7 +42,7 @@ void MenuState::draw()
 {
 	SDL_Rect rectPanel = { 0,0,game->getWindowWidth(),game->getWindowHeight() };
 	backgroundTexture->render(rectPanel);
-	rectPanel = { game->getWindowWidth() / 2 - titleW / 2, titleH / 3, titleW, titleH };
+	rectPanel = { (int)game->getWindowWidth() / 2 - titleW / 2, titleH / 3, titleW, titleH };
 	titleTexture->render(rectPanel);
 
 	/*int x = game->getWindowWidth() / 2 - 250;
